Made folders and files to structure our project.
Made folders and readme's. I think I included all necessary main files and folders. I included my internetradio project, because we are probably going to be doing some similar things as in that project with regards to socket programming and multithreading. By including the radioproject, it will be easier write that code again or even just copy it from there.